split box component; still working on delete and edit

+++ db/delete_donut.sql
@@ -1,2 +1,3 @@
-DELETE FROM dotop WHERE donut_id = $1;
-DELETE FROM donuts WHERE id=$1;
+DELETE FROM dotop WHERE donut_id = $2;
+DELETE FROM donuts WHERE id=$2 AND userid = $1;
+
